+++ macos/Package.swift
@@ -1,0 +1,23 @@
+// swift-tools-version:5.9
+import PackageDescription
+
+let package = Package(
+    name: "media_metadata_plus",
+    platforms: [.macOS(.v10_14)],
+    products: [
+        .library(name: "media_metadata_plus", targets: ["media_metadata_plus"])
+    ],
+    targets: [
+        .target(
+            name: "media_metadata_plus",
+            dependencies: ["MediaMetadataPlusRust"],
+            path: "Classes",
+            resources: [.process("../Resources/PrivacyInfo.xcprivacy")]
+        ),
+        .binaryTarget(
+            name: "MediaMetadataPlusRust",
+            url: "https://github.com/yashas-hm/media-metadata-plus/releases/download/v0.0.0/macos_v0.0.0.xcframework.zip",
+            checksum: "0000000000000000000000000000000000000000000000000000000000000000" // macos
+        )
+    ]
+)